restructure helix into file-level modules
